+++ GoodMorning/GoodMorning/DTO/OpenWeather/CurrentWeatherDTO.swift
@@ -4,34 +4,17 @@
 //
 //  Created by Miro on 2023/07/20.
 //
+import Foundation
 
 struct CurrentWeatherDTO: Decodable {
 
-    let temperature: Temperature
-    let weather: [Weather]
+    let temperature: TemperatureDTO
+    let weather: [WeatherDTO]
 
     enum CodingKeys: String, CodingKey {
 
         case temperature = "main"
         case weather
     }
-
-}
-
-struct Temperature: Decodable {
-
-    let averageTemperature, minimumTemperature, maximumTemperature: Double
-
-    enum CodingKeys: String, CodingKey {
-        case averageTemperature = "temp"
-        case minimumTemperature = "temp_min"
-        case maximumTemperature = "temp_max"
-    }
-
-}
-
-struct Weather: Decodable {
-
-    let icon: String
 
 }
