+++ GoodMorning/GoodMorning/DTO/OpenWeather/TemperatureDTO.swift
@@ -1,0 +1,19 @@
+//
+//  TemperatureDTO.swift
+//  GoodMorning
+//
+//  Created by Miro on 2023/07/24.
+//
+import Foundation
+
+struct TemperatureDTO: Decodable {
+
+    let averageTemperature, minimumTemperature, maximumTemperature: Double
+
+    enum CodingKeys: String, CodingKey {
+        case averageTemperature = "temp"
+        case minimumTemperature = "temp_min"
+        case maximumTemperature = "temp_max"
+    }
+
+}
