+++ GoodMorning/GoodMorning/DTO/OpenWeather/WeatherDTO.swift
@@ -1,0 +1,13 @@
+//
+//  WeatherDTO.swift
+//  GoodMorning
+//
+//  Created by Miro on 2023/07/24.
+//
+import Foundation
+
+struct WeatherDTO: Decodable {
+
+    let icon: String
+
+}
